[Back] AppName - [Front] AdminProductPhotoAction | usePhotoList (#110)
* Update package.json dependencies

* Update package references and shared constants

* Update dependencies in package.json

* Update components and hooks across multiple files

This commit includes several updates across multiple files. In `index.tsx`, unused import statements were removed and new ones were added. The `AdminProductPhotoEdit` component was replaced with `AdminProductPhotoAction` and the `columns` constant was updated accordingly. The `AdminProductPhoto` function was updated to use the `Hub.Product.Get.usePhotoList` hook and the `list` prop was updated. In `Carousel.tsx`, the import statements and `ProductCarousel` function were updated to use the `useAsyncEffect` hook. The `AdminProductDelete` and `AdminProductPost` classes in `Delete.ts` and `Post.ts` respectively, were updated to remove the call to `ProductGet.PhotoListUpdate`. The `ProductGet` class in `Get.ts` was updated to add the `usePhotoList` hook and remove the `PhotoListUpdate` method. A new `AdminProductPhotoAction` component was added in `Action.tsx` which includes the `AdminProductPhotoEdit` component and two `Button` components.