[Back] ProductDeletePhoto | ProductPostMovePhoto | Remove Cover - [Front] DelegateDataGrid | GuidImage | AdminProductPhotoAction | AdminProductPhotoEditCaption | AdminProductPhotoEditDelete | AdminProductPhotoEditReplace | MakeCoverCol | ProductCarousel | usePhoto | useMovePhoto | useGet (#112)
* Update version numbers and disable buttons during loading

* Update Photo model and related code

* Update version numbers and fix code issues

* Update TypeScript files for data grid and product components

This commit includes updates to several TypeScript files related to a data grid component and product-related components in a web application. The changes include modifications to the way certain functions and components are exported and used, updates to the version numbers in the author comments, and changes to the way certain properties are accessed or passed to functions and components. The `manual: true` option has been removed from several `useRequest` calls, suggesting that these requests are now intended to run automatically when their dependencies change. Error handling has been improved with the addition of `onError` callbacks that log errors. Code has been refactored to use destructuring to access properties of objects, improving readability. Interfaces or type definitions have been updated for some functions and components, reflecting changes to the data structures they are expected to work with. Image handling has been updated, with changes to the `GuidImage` component and related functions. Changes have also been made to the way lists of photos are fetched and updated in the `AdminProductPhoto` component and related functions. The changes primarily revolve around the use of hooks and requests in the code, with significant updates to `Patch.ts`, `Post.ts`, `Data.ts`, and `Get.ts`.

* Update dependencies in package.json

* Updated photo handling and UI interactions in various files

Updated and improved photo handling methods in `Delete.cs`, `Post.cs`, `GuidImage.tsx`, `Patch.ts`, and `ObjectStorage.ts`. This includes changes to deletion, moving, and uploading of photos. UI interactions have been enhanced in `Caption.tsx` and `index.tsx` to provide better user feedback during operations. The `ListUpdate` method in `Get.ts` has been marked as deprecated.