Split `Name` field into `Surname` and `Forename`
In this commit, the `Name` field in various classes and functions across multiple files (`Entity.cs`, `Info.tsx`, `Setting.tsx`, `Persona.tsx`, `Get.ts`, `Entity.ts`) has been replaced with `Surname` and `Forename` fields. This change in the data structure splits the name into two separate fields.

In `Entity.cs`, an `Include` statement has been added to include the `User` in the query result when fetching comments. The display of the name in the `Label` and `ToastBody` components in `Info.tsx` and `Setting.tsx` has been updated to display the `Surname` followed by the `Forename`.

In `Get.ts`, the `User` field in the `AdminOrderGet` class and the `Name` field in the `AdminUserGet` class have been replaced with `Forename`. The `name` variable in the `OrderGet` class is now assigned the value of `user.Forename` instead of `user.Name`.

Finally, in `Entity.ts`, the `AdminUserEntity` class has been updated to version `0.2.0` and the `Name` field has been replaced with `Surname` and `Forename`.